feat: overhaul extension to work as a kysely driver to enable plugins and a more fluent qpi
Remove the $kyselyQuery and $kyselyExecute methods. Instead, allow kysely queries to be run in the
usual manner with .execute(), making the query api more intuitive. To accomplish this, the extension
now includes a custom driver for kysely, exposed via the extension's constructor.

BREAKING CHANGE: Remove the $kyselyQuery and $kyselyExecute methods.
fix #29